End-to-end tests were relocated to the <tests> directory

+++ tests/end_to_end/src/end_to_end.cpp
@@ -5,10 +5,10 @@
 
 int main ()
 {
-    auto cin_iter = std::istream_iterator<int> (std::cin);
+    auto cin_iter = std::istream_iterator<double> (std::cin);
     size_t size = *cin_iter;
 
-    yLab::Matrix<int> matrix {size, size, ++cin_iter, std::istream_iterator<int>()};
+    yLab::Matrix<double> matrix {size, size, ++cin_iter, std::istream_iterator<double>()};
     std::cout << matrix.determinant() << std::endl;
     
     return 0;
